[Refactor]: ViewModel, I/O, Protocol 분리
[Refactor]: 기존 ViewModel에서 I/O와 프로토콜 분리
