+++ tls427wodnr/AdvanceProject0509/AdvanceProject0509/Presentation/BookDetail/ViewModel/BookDetailBottomSheetViewModelProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  BookDetailBottomSheetViewModelProtocol.swift
+//  AdvanceProject0509
+//
+//  Created by tlswo on 5/15/25.
+//
+
+// MARK: - BookDetailBottomSheetViewModelProtocol
+
+protocol BookDetailBottomSheetViewModelProtocol {
+    func transform(input: BookDetailBottomSheetViewModelInput) -> BookDetailBottomSheetViewModelOutput
+}
