+++ tls427wodnr/AdvanceProject0509/AdvanceProject0509/Presentation/BookList/ViewModel/BookListViewModelProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  BookListViewModelProtocol.swift
+//  AdvanceProject0509
+//
+//  Created by tlswo on 5/15/25.
+//
+
+// MARK: - BookListViewModelProtocol
+
+protocol BookListViewModelProtocol {
+    func transform(input: BookListViewModelInput) -> BookListViewModelOutput
+}
