+++ tls427wodnr/AdvanceProject0509/AdvanceProject0509/Presentation/RecentBook/ViewModel/RecentBookListViewModelProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  RecentBookListViewModelProtocol.swift
+//  AdvanceProject0509
+//
+//  Created by tlswo on 5/15/25.
+//
+
+// MARK: - RecentBookListViewModelProtocol
+
+protocol RecentBookListViewModelProtocol {
+    func transform(input: RecentBookListViewModelInput) -> RecentBookListViewModelOutput
+}
