+++ tls427wodnr/AdvanceProject0509/AdvanceProject0509/Presentation/Search/ViewModel/SearchViewModel.swift
@@ -9,27 +9,7 @@ import Foundation
 import RxSwift
 import RxCocoa
 
-// MARK: - Input
-
-struct SearchViewModelInput {
-    let query: Observable<String>
-    let loadMoreTrigger: Observable<Void>
-}
-
-// MARK: - Output
-
-struct SearchViewModelOutput {
-    let books: Driver<[BookItem]>
-    let error: Signal<Error>
-}
-
-// MARK: - Protocol
-
-protocol SearchViewModelProtocol {
-    func transform(input: SearchViewModelInput) -> SearchViewModelOutput
-}
-
-// MARK: - ViewModel
+// MARK: - SearchViewModel
 
 final class SearchViewModel: SearchViewModelProtocol {
