+++ tls427wodnr/AdvanceProject0509/AdvanceProject0509/Presentation/Search/ViewModel/SearchViewModelProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  SearchViewModelProtocol.swift
+//  AdvanceProject0509
+//
+//  Created by tlswo on 5/15/25.
+//
+
+// MARK: - SearchViewModelProtocol
+
+protocol SearchViewModelProtocol {
+    func transform(input: SearchViewModelInput) -> SearchViewModelOutput
+}
